pyproject.toml
[project]
name = "audio-driven-avatar-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "librosa>=0.11.0",
    "mediapipe>=0.10.35",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "python-dotenv>=1.2.2",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.8.0",
]
